ext65 refactor
