+++ keyboards/aeboards/ext65/config.h
@@ -1,0 +1,1 @@
+#include "config_common.h"
