+++ keyboards/aeboards/ext65/ext65.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "quantum.h"
+
+#ifdef KEYBOARD_aeboards_ext65_rev1
+    #include "rev1.h"
+#endif
+
+#ifdef KEYBOARD_aeboards_ext65_rev2
+    #include "rev2.h"
+#endif
