--- keyboards/aeboards/ext65/rev1/keymaps/via/readme.md
@@ -1,2 +1,0 @@
-# The VIA Ext65 Layout
-
